Add --skipdotnet flag to install script

+++ shrimpcast/Linux/install.sh
@@ -9,6 +9,7 @@ handle_error() {
 APPLICATION=""
 NGINX=""
 RESTREAMER=false
+SKIP_DOTNET=false
 
 if [ "$(id -u)" -ne 0 ]; then
     echo "This script requires root privileges. Please run this script as root (sudo -i)."
@@ -27,6 +28,10 @@ while [[ $# -gt 0 ]]; do
       RESTREAMER=true
       shift
       ;;
+    --skipdotnet)  # New parameter for skipping .NET installation
+      SKIP_DOTNET=true
+      shift
+      ;;
     *)
       echo "Invalid option: $1" >&2
       exit 1
@@ -40,6 +45,14 @@ if [ -z "$DOMAIN" ]; then
   exit 1
 fi
 
+# Logic for skipping .NET installation if --skipdotnet is provided
+if [ "$SKIP_DOTNET" = true ]; then
+  echo "Skipping .NET installation."
+else
+  # Your logic for .NET installation goes here
+  echo "Proceeding with .NET installation."
+fi
+
 # Output the parsed parameters
 echo "Using domain: $DOMAIN"
 if [ "$RESTREAMER" = true ]; then
@@ -47,7 +60,6 @@ if [ "$RESTREAMER" = true ]; then
 else
   echo "Restreamer option not enabled."
 fi
-
 
 # Check if UFW is installed
 if command -v ufw &> /dev/null
@@ -108,21 +120,23 @@ rm -rf /root/nginx_tmp
 # Get Ubuntu version
 declare repo_version=$(if command -v lsb_release &> /dev/null; then lsb_release -r -s; else grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"'; fi)
 
-# Download Microsoft signing key and repository
-wget https://packages.microsoft.com/config/ubuntu/$repo_version/packages-microsoft-prod.deb -O /root/packages-microsoft-prod.deb || handle_error "Failed to download Microsoft signing key and repository"
-
-# Install Microsoft signing key and repository
-dpkg -i /root/packages-microsoft-prod.deb || handle_error "Failed to install Microsoft signing key and repository"
-
-# Clean up
-rm /root/packages-microsoft-prod.deb || handle_error "Failed to remove packages-microsoft-prod.deb"
-
-# Update packages
-apt update || handle_error "Failed to update packages"
-
-# Install .NET SDK 8.0
-apt-get update && \
-  apt-get install -y dotnet-sdk-8.0 || handle_error "Failed to install .NET SDK 8.0"
+if [ "$SKIP_DOTNET" = false ]; then
+    # Download Microsoft signing key and repository
+    wget https://packages.microsoft.com/config/ubuntu/$repo_version/packages-microsoft-prod.deb -O /root/packages-microsoft-prod.deb || handle_error "Failed to download Microsoft signing key and repository"
+    
+    # Install Microsoft signing key and repository
+    dpkg -i /root/packages-microsoft-prod.deb || handle_error "Failed to install Microsoft signing key and repository"
+    
+    # Clean up
+    rm /root/packages-microsoft-prod.deb || handle_error "Failed to remove packages-microsoft-prod.deb"
+    
+    # Update packages
+    apt update || handle_error "Failed to update packages"
+    
+    # Install .NET SDK 8.0
+    apt-get update && \
+      apt-get install -y dotnet-sdk-8.0 || handle_error "Failed to install .NET SDK 8.0"
+fi
 
 dotnet dev-certs https
